rivreader/requirements.txt
PyPDF2==3.0.1
ebooklib==0.18
beautifulsoup4==4.12.2
anthropic==0.32.0
ttkthemes==3.2.2
python-docx==0.8.11
striprtf==0.0.21
html2text==2020.1.16